Ultrasonic Barrier Detector/requirements.txt
autopep8==1.5.7
cycler==0.10.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.3.2
pycodestyle==2.8.0
pyparsing==2.4.7
pyserial==3.5
python-dateutil==2.8.2
six==1.16.0
toml==0.10.2
